trigger/requirements.txt
requests
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-jaeger
deprecated
opentelemetry-instrumentation-requests 
